Move remaining utility functionality out of tuple.hpp into tuple_utility.hpp
